+++ _goldens/test/_files/queries.dart
@@ -41,6 +41,9 @@ class QueriesComponent {
 
   @ViewChildren(AnotherDirective)
   List<AnotherDirective> usingTypeFromField;
+
+  @ViewChild('q3')
+  AnotherDirective nestedViewChild;
 }
 
 @Component(